Disable autocomplete on uid fields
We have a number of fields where the user has to enter a unique
identifier, having autocomplete enabled on fields like these increases
the chance of errors at the cost of easily entering repeated uids, which
feels like a good trade off.

If a users enters an incorrect uids, the chances of doing so again also
increase as the browser will store the incorrect uid for
autocompletion.