Fix: Service support contact permissions
During a product review we noticed that as a service support users we
couldn't edit contacts, our product owner felt this group of users
should be able to edit contacts, particularly the chair of governors
once added, this work fixes this.

The cause was the `has_role?` method does not include service support
users, the solution was a early return for service support users,
generally speaking, service support users can do 'anything'.