Add advisory board date in range scope
We want to tidy up our exports and the first step was to identify the
main attributes that define which projects are included in the exports.

One of these is the advisory board date and we want a scope that can
include projects where the date falls within a range, rather than the
existing month.

Later work could remove the single month scope, instead passing a single
month as a range, but that is not part of this work.