Move scoping of project exports
The 'by month' export has a different criteria for which projects to
include than the other exports.

We want to offer all exports with a date range and want to utilise the
`significant_date_in_range` scope but without the `in_progress` and
`confirmed` scopes pre applied.

To do so we move those pre scopes to where they get used to render the
'by month' views and exports, leaving the scope focused only on the date
range and available for use in other exports.