Add date picker options helper
We will be adding more date pickers and want a simple way to build the
list of options for them.

This helper does just that, the ids are Date objects and the names are
GOV.UK short dates e.g. "Jan 2024"

From dates are always the 1st and to dates are always the last day of
the month.