Remove trust validations from the Project model
Background: We have an issue on Production where projects keep becoming invalid.
The UKPRN entered for the incoming trust, which was valid when the project was
created, is being changed upstream in the Academies DB. That change is then
being ingested into the Academies API, and when we try to load an existing
project, the UKPRN is not found and it significantly breaks the application
for users.

Because the Project model is "at rest", we do not think it needs validations.
Any Project creation or editing is done via FormObjects, which have their own
validations. We do not need to be validating the Projects on load into listing
pages, view pages or exports.

To start, remove the Trust validations from the Project so that when a project
is loaded with an invalid incoming or outgoing UKPRN, it does not raise an
exception which we cannot handle.

This also involves removing an expectation from the Academies Pre-Fetcher spec,
as the pre-fetcher will no longer call `get_trust` to check a trust registered
to a project actually exists.