Return a fake Trust object if a Trust is not found on Academies API
Background: We have an issue on Production where projects keep becoming invalid.
The UKPRN entered for the incoming trust, which was valid when the project was
created, is being changed upstream in the Academies DB. That change is then
being ingested into the Academies API, and when we try to load an existing
project, the UKPRN is not found and it significantly breaks the application
for users.

Because the Project model is "at rest", we do not think it needs validations.
Any Project creation or editing is done via FormObjects, which have their own
validations. We do not need to be validating the Projects on load into listing
pages, view pages or exports.

If a UKPRN is invalid, load up a fake Trust object. This means the listing, view
pages and exports will not break. Show the error message as the trust "name"
instead, to indicate there is an issue with the UKPRN.