Create export form objects
We are making the csv exports more conventional and for users to supply
a date range for the exported projects.

We start with the RPA, SUG and FA letters export, first creating a base
class that we can share across all the exports later.

The form accepts a from and to date and returns the appropriate csv
export using the existing service and scopes.