Replace the RPA, SUG and FA letters export
We treat each export as a resource and add new and create paths for
them, we then add conventional controllers and views to allow users to
supply date ranges and create a new 'export'.

We do not use the GOV.UK `select` elements as we do not want to show
validation errors inline as this would break the date picker.