Do not raise in the Academies API prefetcher encounters an unknown trust UKPRN
Background: We have an issue on Production where projects keep becoming invalid.
The UKPRN entered for the incoming trust, which was valid when the project was
created, is being changed upstream in the Academies DB. That change is then
being ingested into the Academies API, and when we try to load an existing
project, the UKPRN is not found and it significantly breaks the application
for users.

Because the Project model is "at rest", we do not think it needs validations.
Any Project creation or editing is done via FormObjects, which have their own
validations. We do not need to be validating the Projects on load into listing
pages, view pages or exports.

If the Academies API prefetcher encounters a UKPRN which is not valid (not on
the Academies API), do not raise - we cannot handle an exception gracefully
here. Instead return nothing and allow the page to still be loaded.