Add MS Application Insights connection string
In order for the Application Insights browser telemetry to work we need
to pass the connection string from the environment to the JS.

We'll output the string to a meta tag so that it can be read by the JS
later.

The string would have been made public in any approach we take.

We only need to show the string if all the conditions for running
analytics are present, these are:

- the string is present in the environment
- user has opted in to option cookies
- the application is running in one of the production environments (dev,
  test, prod)