Enable MS Application Insights in the browser
This is the final step to enable Application Insights in the browser.

We create a JS file to load and initialise the code and conditional add
the result to the page if the user has allowed the tracking.

We have to modify the content security policy to allow the connection to
Application Insights.