Add SignificantDateHistoryReasons
We will need to start recording a reason for a chaange to Significant Dates on
projects. These reasons will be taken from a list of predefined reasons in a
form. TO begin, we add a SignificantDateHistoryReason model, which sits between
the SignificantDateHistory record and its Note.

To ensure we do not break the existing SignificantDateHistory functionality, we
need ot preserve the connection between SignificantDateHistory and Note. Once
the new SignificantDateHistoryReasons are in place, we can do a data migration
to break these connections and re-join the existing SignificantDateHistory
records to their Notes via a SignificantDateHistoryReason record.