Significant date service refactor
We now have the concept of a 'reason' the significant date has changed
and need to update this class which creates all the required objects to
include the reason.

We have also created an enum on the `reason_type` to ensure only valid
types can be used, we only have two at this point:

- the stakeholder kick off task updated the date
- something else changed the date that we don't have a type for at this
  point

Right now there will only be one single reason, but later work will
introduce more.