SignificantDateHistoryReason data migration
We have added this model between `SignificantDateHistory` and it's note
and we need to add a `SignificantDateHistoryReason` to existing data.

This migration fetches every project regardless of state and then
fetches every date history.

It creates a new reason with type:

- `stakeholder_kick_off` when the note body is appropriate
- `legacy_reason` for all other cases

The reason creation date is copied from the date history and the
existing note linked.

As the existing model can only have a single note, all existing data can
only have a single reason, multiple reasons will not be possible until
the UI for that is built.

We also set the SignificantDateHistory user association to that of the
note already attached.

We are not removing the link between the data history and the note in
this migration, that will come later.

This migration is intended to be run in production once deployed, we
have console access and then be deleted.