Add user association to SignificantDateHistory
Date histories can now have multiple 'reasons' with attached notes so
there is no longer a one to one relationship between a date history and
it's note, which we used to infer who created the date change.

We'll store the user on the date change itself from this point forward.