Update the new date history form
Changing the significant date for a project will be a two step process,
with the first step to collect the new date.

This work updates the form object and form views with the new content
and removing the note, which will be collected on the second step to
follow.

We remove the feature spec as will will create a new one that exercises
both steps.

As the form object no longer saves anything, the application is broken
at this point!