Don't include `linting` gems in the Dockerised application
We have had an issue twice now where the latest version of StandardRB v1.37.0
has caused the deploying container to crash on Azure. Aside from attempting to
figure out why this happens on deployments but not locally, we do not actually
need to include Standard (and ERBLint) in our Dockerised version.

Exclude the gems in the `linting` group when building the application in Docker.