Add significant date earlier form
When the significant date is changed to an earlier one, we need to
collect the reason why. This form object will handle that.

The reasons for conversions and transfers are the same.

At least one reason must be selected and a note must be added for each
selected reason.

Later we will add the 'later' version of this form object, that will
likely share behaviour.