Add significant date later form
The later form works the same as the earlier form except it has to
handle two different list of reasons and way more reasons.

To do this, we move all the shared logic from the existing earlier form
into a base class and add the later form.

As the later form has so many reasons, we use some meta programming to
add the attributes and validations to the class based on the list of
reasons we need to have anyway.

We also add all of the available reasons to the
`SignificantDateHistoryReason` model and locales to get the validation
working.