Prefer a Contact::Establishment record for establishment contact
We have had an issue where Headteacher contacts are not being selected for the
'school contact' in exports. Because not many people use the "Main school
contact" checkbox, the first `school_or_academy` contact was being picked as the
main establishment contact in exports; however this contact is not usually
correct.

If there is a Contact::Establishment in the array of `school_or_academy`
contacts, prefer that contact over any other.

(`in_order_of` is new in Rails 7!)