(Refactor) Set the MP's address to always be House of Commons
An MPs parliamentary office address is always the House of Commons, and that is
the only address we're interested in. We do not need to get their constituency
office address.

Remove some of the complexity from the MP retrieval code by only retrieving the
parliamentary office details (we are specifically interested in the MP's
parliamentary office email)