Allow users to select from ALL project contacts for a Project Main Contact
We had a bug where users were not able to select an ingested Head Teacher
contact as a "Main Contact" for a project.

Because the ingested Head teacher contacts are `Contact::Establishment` records
and not linked to the project via an association, they do not show up when
you call Project.contacts.

We have added a new method, Project.all_contacts, which uses the ContactsFetcher
to return *all* contacts, including Contact::Establishment and
Contact::DirectorOfChildServices. Use this method in the view, and therefore
allow users to select them as a Main contact for a project (they are not likely
to choose the Director of Child Services though!)