Add form object for DAO Revocation
This form object will allow users to 'create' a new Directive Academy
Order (DAO) revocation. The data model is an association on Project and
only Conversion projects with a DAO can have them revoked.

The designs for this call for a 'check your answers' flow and a custom
'confirmation' that must be completed by the user but collects no stored
data.

We'll use validation contexts to validate the steps as a user moves
through the flow.

As the form object also models the steps of the flow, we have some
additional helper class methods.

We also collect a date, so we have to validate the three field date.