Change project state on DAO revocaiton
When a project has it's DAO revoked it is 'finished' in the sense there
will be no more work on it.

We store this on the `state` attributes of Projects, although DAO only
applies to conversions, state is global to all project types.

We will add a number of behaviours based on this new state, but that
does not form part of this work.