User can revoke a DAO
This work exposes the new 'DAO revocation' UI to users. It follows the
same pattern as the 'completing a project' UI.

We show a link and some brief guidance at the bottom of the task list
of appropriate projects. We also add a link to the side navigation. The
logic for showing the links in the side bar is the same as that of
showing the guidance and links.