Refactor ByMonthProjectFetcherService
These two pairs of methods should return the same set of projects with
the same sort order, which means we can DRY this up.

We also add a new scope for this service: `active`. This only includes
projects that are not completed or have the DAO revoked.

A single month can be represented by passing two dates in the same month
to the ranged fetcher. Although these dates are always the 1st, we felt
the code should handle any case so we take steps to endure the from date
is the 1st while the to date is the end of the month.

We then have to ensure that even on routes that pass only a single date,
we turn that into a from and to date, at the same time we try to ensure
we are passing a Date object as much as possible, casting to strings as
needed rather than passing strings about.

We added the DateRangable controller concern for the common behaviour
between the conversion and transfer controllers.