Include the state of DAO in export
We think the users of the pre conversion grant export will want to see
DAO projects, but there would be no way to identify them.

This work adds a 'Directive academy order revoked' column to this export
so that users can identify any that have been revoked.

The value will be one of:

- Not applicable, when the project has no DAO
- 'no', when the project has a DAO but it has not been revoked
- date, the date on which the DAO revocation decision was made, this
  value is capture during the revocation steps