Update optional cookies cookie
During the June 2024 IT Health check it was raised that our
`ACCEPT_OPTIONAL_COOKIES` should be httponly.

This change will not take effect for users who already have the cookie
set, until they delete the existing cookie or it expires.

Clearing cookies for the application will set the new cookie and is
recommended.