Update Google tag manager
Around June 13, possibly when we added MS Application Insights, Google
Tag Manager stopped running on the application.

This is an attempt to fix that.

We make the approaches consistent by outputting the Tag Manager ID as a
meta tag and having the Tag Manager script in our application
JavaScripts, this is how we handle Application Insights.

We only load the meta tag and script if optional cookies are accepted.

This work also attempts to simplify the content security policy, we
default to allowing self and then only the `connect-src` to fetch the
third party scripts for Tag Manager and Application Insights.

We removed the nonce generator as the nonce is never output to the page,
so cannot work? (we think?) and is no longer needed anyway.

We are not 100% confident this will allow Tag Manager to function, but
we think the parts we use will and only 'Custom JavaScript Variables'
will fail, according to out take on this:

https://developers.google.com/tag-platform/security/guides/csp

As we do not have access to Tag Manager and it only runs in production,
there is no way to verify until we ship this.