Fix cookie form
We keep seeing deployments fail with this error:

```
/srv/app/app/forms/cookies_form.rb:7:in `<class:CookiesForm>': uninitialized constant CookiesForm::OpenStruct (NameError)
```

So far this only seems to happen when we update the Standardrb gem and
backing that out has consistently fixed the issue, although it could
just be a coincidence!

We cannot reproduce the issue locally either.

If the failure is this OpenStruct in CookiesForm the only thing unusual
about it is the assignment to a const, there is really no need to do
this, so let's try it without and see how we get on.