Add a form, view and routes for new projects
Instead of showing four buttons to add projects, we want to show a form
of radio buttons, this will give us more opportunity to guide users in
which type is the right one for their needs.

We can redirect based on their response to the existing new forms we
already have.