Add notable association to Note
We always associate notes to a project and a user, we also want to
associate them to other entities, right now this is a
SignificantDateHistoryReason, which used to be SignificantDateHistory.

This work changes the association on Note to it's parent
SignificantDateHistoryReason so it is polymorphic, i.e. it can have
different classes of parents - this will allow us to association to a
DaoRevocation in later work.

Right now ALL notes with a significant_date_history_reason_id will
become notable_type "SignificantDateHistoryReason" - whilst the
migration includes what looks like a data migration, it is not one as
notable_type is internal structure for Rails and not data from the
application.

There is a bunch of leftover code from the old SignificantDateHistory,
but we won't remove that here, but we will remove it!