Add DAO Revocation Reasons model
We currently store the 'reasons' for a DAO revocation as boolean column
values, but we now want to also store free text to explain the reason
along with it.

We already do the same thing for SignificantDateHistory so it feels
right to do it here as well.

Using a simple join model DaoRevocationReason sits between the
DaoRevocation and a Note.

The advantage of this is to collect ALL notes relating to a project as
one model and in one table. The disadvantage is the additional
complexity.