Configure the Rails cache to use Redis
Whilst we do not use the Rails cache for caching, we want to use it to
store temporary values on our stepped forms.

This work configures the cache to use Redis and enable it in all
environments.

In development caching can be toggled on and off and we have retained
this behaviour.

You will need to be running Redis locally and have the REDIS_URL
variable set in .env.development.local and .env.test.local from this
point forwards.