Switch DAO revocation stepped form to cache
We are adding long form responses to the form in later work and the 4kb
cookie limit for the cookie store could potentially be reached.

To avoid the issue we have configured the Rails cache, backed by Redis
so we can store the values there.

We have a one hour expiry on the values so they are deleted after that
time.

We couldn't use the preferred option of moving the session store to the
cache as the OmniAuth CSRF protection is not compatible with the cache
store.