Allow users to edit the "Handover to RCS" field on projects
The RCS team have reported that users creating projects, intending to hand them
over to RCS, are not filling out the project creation form correctly and
instead keeping the projects assigned to themselves. The RCS team are then
having to spend time every day informing users how to reassign projects
via the Internal Contacts tab.

In order to try to mitigate this, make the "Handover to Regional casework
services" field on the Edit project form editable. This field is not actually
an attribute on the project, it adjusts who and to which team a project
is assigned to.

The user can now assign the project to RCS via the edit project form. This does
however mean they will no longer be able to edit the project, as the project
is no longer assigned to them (!). But this should hopefully reduce confusion
and give the RCS team less daily work.