Fix notable migration
We got this migration through dev and test but it failed in prod as
there are 'deleted' projects which are excluded by a `default_scope`.

This means the notes for these projects return nil for their project
making them invalid and so not saving.

Our plan is to rollback the migrations in dev and test and redeploy this
fix.