Make Contact::Parliament a "real" contact type
We're going to have to store the MP details in our app because the Members
API is too inefficient to use on the fly. First step, turn the transient
Contact::Parliament model into a "real" ActiveRecord model so we have somewhere
to store our MP details.

Note that the `parliamentary_constituency` has to be a string because we're
going to link Establishments and MPs together by constituency. The constituencies
on GIAS do have IDs, but they're not the same ID as the ones used on the Members
API. So we have to do the matching by constituency name.