Add "project main contact" details to RPA/SUG report
A user reported that they were adding a contact as the "Project main contact"
but that contact was not being output to the RPA/SUG report. Despite the
"select the main contact for a project" task
indicating that "this person will receive funding letters", the main contact
for a project was not actually being included in the report.

As a temporary fix, add this contact into the RPA/SUG report. When we rework the
contacts pages, this commit will become redundant, as users will be able to
choose the contacts output to the RPA/SUG report.