Update ByTrustProjectFetcherService
We want to include projects that are forming a multi academy trust in
the 'by trust project view'.

This is made difficult by the fact there is no model for a form a MAT
trust, only two attributes on relevant projects:

- new_trust_reference_number
- new_trust_name

This work solves this by building the two sets and then combining them
before sorting.

The service already collects the trusts and groups projects so this is
mostly refactoring.

For the form a MAT projects we have to load all appropriate projects
(those with the two attributes) and build a unique list of 'form a
MAT trusts' from that by assuming the trust reference number is unique
along with the name provided (the name is more of a human thing and is
in theory validated at the point it is added to be the same across
matching records).

We make fake `Api::AcademiesApi::Trust` objects from each record which
we can use to get the counts of projects that match, fake objects is how
we solve this issue elsewhere in the application.

This hangs off the idea that the trust reference number is unique and
fetching the first one is good enough with the call to
`find_by_new_trust_reference_number`.

At this point the application is broken as the view that renders this
relies on the UKPRN which for form a MAT trust will be nil, we'll fix
this next.