Update by trust view
Now that the service that backs this view can include both 'regular'
projects and 'form a multi academy trust' projects we need to make some
changes to support this.

The index view of all trusts with active projects gets logic to render
either case.

The show view gets a completely new action and route because it is by
trust reference number instead of UKPRN. We used the same view file as
there is no difference.

We simplify so specs here using more generic API mocking as we don't
need as specific coverage.