Add ProjectGroup model
Conversions and transfers are 'grouped' together at advisory board to
join the same trust and doing so opens the door to funding under new
policies.

We need to support this grouping of projects, but will not originate the
groups as this is done pre-advisory board (so in the Prepare
applications).

A project can only belong to one group, but a trust could have multiple
groups over time.

Right now a group will always relate to a Trust, for now we will capture
the incoming trust UKPRN to make this connection.

This work sets up a basic model to handle this.