Conversions collect group reference number
We now have ProjectGroup which represents a group of projects that went
to advisory board together for the same trust and so are entitled to
additional funding.

This work takes the steps to collect the 'group reference number' which
is the grouping attributes in the Prepare application.

Right now, users have to provide the group reference number, once the
two applications are connected, it can simply be passed.

We validate the group reference number:

- it must be in the correct format when provided
- when the ProjectGroup already exists, the `incoming_trust_ukprn` must
  match that of the ProjectGroup `trust_ukprn` i.e. the conversion must
  be joining the same trust.

We add to the base `create_project_form` because both conversions and
transfers have this behaviour, it will be added to transfers next.

Along the way we improved a spec helper used to fill in the create
conversion project form, extracting it for reuse.