Transfers collect group reference number
Just like conversions, transfers can be grouped together for advisory
board to receive additional funding.

This work collects the group reference number, validates it and makes
the grouping.

As both conversions and transfers have this behaviour, much of the code
is already in place from the conversions work and this is a repeat on
transfers.