(Fix) Allow all Contact types to be a "project main contact"
When testing PR #1780 in development, we realised that any Contact type other
than Contact::Project, when selected for the "main contact for the project"
(Project.main_contact), was not showing in the exports.

This was because the contact type for project main contact was set to
Contact::Project, but the "Choose a main contact for the project" task allows
*any* contact type to be selected.

Additionally, when fixing this issue we realised that he Chair of Governors
contact association was only on the Conversion Project type, when we allow
Transfer projects to choose a Chair of Governors as well.

Fix both of these by moving the associations into the base Contact and Project
classes.