Group reference number editing
Now that projects can be grouped, we need to be able to edit the
grouping in-case of errors.

This work does so following the existing pattern, utilising the edit
forms.

Of note here is that we make sure to declare each attribute type as
comparing them later should be consistent e.g. UKPRN are always
integers and we compare them.

We also have to support removing the group association which the user
achieves by editing the group reference number to an empty string.

There is a little duplication of the validator here, but we will fix
that in a separate commit.