Add project group pre fetching service
Like the one we have for projects, we can improve performance by
prefetching the trusts for a given set of `ProjectGroup`.

This service does that, taking a set of groups, fetching the trusts
objects and setting them on to the groups.

The testing is a little clunky as all we can really do is check that the
right number of API calls are made, so:

- when prefetched we expect the bulk `get_trusts` endpoint to be called
- when not prefetched we expect the bulk endpoint to be called, which
  fails then fall back to the `get_trust` endpoint a number of times to
  fetch each trust.

We fail silently (apart from sending an event to MS Application
Insights) as the fall back will set a fake trust object with the error
as the name, which is our approach elsewhere.

Whilst pre fetching the grouped projects would be ideal, to do so would
require some significant refactoring of the objects as they are
associations, we decided not to do this and see how the application
performs in production before attempting that work.