Add key contacts model
Key contacts are those used by data consumers to send important
communications to those involved in projects.

They fill DfE defined roles in the project:

- headteacher
- chair of governors
- incoming trust CEO
- outgoing trust CEO
- director of children services
- member of parliament

These roles are not the contact job title.

We want to encapsulate these into a model that sits outside of Project
(to keep Project as clean as we can).

We settled on `key_contacts` which is a simple joining model between the
project and the contacts.

Here we are only adding headteacher, but the others will follow.

We have no referential integrity here as:

- a project will destroy its own key_contacts
- we want a contact to be deleted even if referenced from a key_contact,
  which will then return `nil`, which is the expected behaviour

There is also no need for polymorphic associations here as Contact use
single table inheritance so that Rails can identify the class of the
contact that is returned on the association.

This approach is a similar to `TasksData` which we use to keep all the
task related attributes out of the project model and table.