Add FormObjects to create specific contact types
We want to "funnel" particular Contacts to have specific roles and categories,
e.g. a school contact who is a Headteacher, or a trust contact who is a CEO.

To begin, we add FormObjects to create Contact::Project records with
fixed job titles and organisation names. E.g. a Headteacher for a school whose
organisation name is the name of that school.