Add email and phone to contacts
The design for the new 'confirm the headteacher' task calls for us to
show the name of the contact along with the email address and phone
number separated by a comma.

This method will be used later to display this information.

We decided not to add a presenter at this stage as this is the only call
for something specific.