Add confirm headteacher contact task
We are going to ask users to specifically identify the headteacher
contact in an effort to improve the quality of the data being provided
from the application.

We run the generator to create the scaffold of the task.

The task sets the association on the recently added `KeyContacts` class
rather than the `Project` class.

Only contacts from the `school_or_academy` category are used to populate
the available contacts, although this is not enforced beyond this.