New Contact Creation Flow
Apoligies for the large commit but it wasn't possible to do this in a smaller
shot.

We want to ensure users create contacts with a very specific job title, for
example "Headteacher" for whoever is the "headteacher-like" person at a school,
or "CEO" for the person in charge of a trust. Because we previously allowed
users to type anything into the Job title/role field for contacts, there was
confusion for end users (usually export users) over which users have which
roles, and who to send certain types of paperwork to.

This commit steers users to create contacts of a certain role and predetermines
their job title. For example, the Headteacher's job title will always be
"Headteacher".

Users can still create "other" contacts with different job titles in all
categories, but we hope this creation flow will encourage them to always
create at least *some* contacts with the expected job titles.