Add confirm chair of governors task
Just like the headteacher version, this new task is more explicit about
what it is asking for so that downstream data users can be assured of
the quality of the data.

There is an existing 'Add chair of governors' contact details' task that
will be removed soon and used to migrate any existing data.