Data migration for chair of governors
We want to copy any chair of governors contacts that have been added
with the old task to the new one.

The test setup is a bit unexpected here because of how the
`chair_of_governors_contact` association works, we have to create and
associate objects in a specific order.

This data migration should be deleted once run in production.