Remove headteacher column from export
Our new 'contact tasks' are about to be deployed but we realised that we
are about to enter the largest openers month of the year, September and
doing so will introduce gaps in the data that downstream users rely on.

On reflection, we decided that it would be better to leave this export
as is until after September.