Add DfE Persons API client
A new API has been added internally at DfE so we can get the member of
parliament, the 'Persons API':

https://s184d01-tramsapicdnendpoint-personsapi-a3d9cpcwahc6dnh0.z01.azurefd.net/swagger/index.html

We'll use this as a replacement for the Members API.

This work introduces a new API client for the Persons API and fetches a
single member object from a supplied constituency name.

We decided to drop the `api` suffix as we are inside the `api` namespace
already.