Switch to the Persons API for MP
Whilst the Persons API is not in production yet, we can get everything
ready for when it is.

This work switches the Members API to the Persons API but leaves it
unused.

We do use it on the hidden `project/id/mp` view just to validate things
are working in development.

At this point test coverage will fall below 100% as we are no longer
calling the Members API code, but we will remove that next.