Allow users to select the GIAS-imported Headteachers in the Headteacher task
After deploying the latest changes, users were confused as to why they were unable
to select the GIAS-imported Headteachers in the new Headteacher task. We had
previously decided the data quality in these contacts was not good enough and
hoped users would be able to enter their own Headteacher details. However
some users still want to be able to use this data.

As a fix, we have decided to allow users to choose the GIAS-imported Headteachers,
despite them being of dubious data quality.