Data migration to copy the establishment contacts
We recently introduced new tasks to confirm the headteacher contact.

In the past we have imported the headteacher from GIAS data to a special
shared 'Contact::Establishment`. This is no longer the case as we were
told that, for conversions at least, the contact details in GIAS are not
reliable.

With the new tasks released, users were confused as they could see the
headteacher contact but not use it for the task.

As we are no longer importing the GIAS data and want all users to
confirm the headteacher contact we will copy the establishment contacts
to editable project contacts, which will allow users to:

- select the imported contact if is correct
- edit the contact if it is incorrect

When run, this data migration will appear to create duplicates, but we
will stop showing `Contact::Establishment` contacts to users next so
they will not know.

Once we are happy, the `Contact::Establishment` can be deleted.